+++ tests/testthat/test-save-app.R
@@ -1,5 +1,49 @@
 require("shiny", quietly = TRUE, character.only = TRUE)
 
+# App from: https://github.com/rstudio/shinytest2/issues/303#issuecomment-1377950984
+test_that("Make sure global vars are set - issue303", {
+  # Only run on CI. CRAN requires package to be installed to run `callr::rscript()
+  skip_if(!on_ci(), "Only run on CI")
+
+  # Run the script in a global environment that does not polute this global environment
+  p <- callr::rscript("scripts/issue_303.R", show = FALSE)
+  expect_equal(
+    any(grepl("object 'foo' not found", p$stdout, fixed = TRUE)),
+    FALSE
+  )
+  expect_equal(
+    any(grepl("object 'foo3' not found", p$stdout, fixed = TRUE)),
+    FALSE
+  )
+})
+
+# App from: https://github.com/rstudio/shinytest2/pull/307#issuecomment-1381391531
+test_that("Make sure global vars are set - pr307", {
+  # Only run on CI. CRAN requires package to be installed to run `callr::rscript()
+  skip_if(!on_ci(), "Only run on CI")
+
+  # Run the script in a global environment that does not polute this global environment
+  p <- callr::rscript("scripts/pr_307.R", show = FALSE)
+  expect_equal(
+    any(grepl("could not find function", p$stdout, fixed = TRUE)),
+    FALSE
+  )
+})
+
+# App from: https://github.com/rstudio/shinytest2/pull/307#issuecomment-1381391531
+test_that("Make sure global vars are set - issue295", {
+  # Only run on CI. CRAN requires package to be installed to run `callr::rscript()
+  skip_if(!on_ci(), "Only run on CI")
+
+  # Run the script in a global environment that does not polute this global environment
+  p <- callr::rscript("scripts/issue_295.R", show = FALSE)
+  expect_equal(
+    any(grepl("could not find function", p$stdout, fixed = TRUE)),
+    FALSE
+  )
+})
+
+
 test_that("Saving an app has the right context", {
   n <- 5
   shiny_app <- shinyApp(